requirements.txt
requests
numpy
pandas
plotly
pymongo
python-dotenv
beautifulsoup4
streamlit
selenium
streamlit-authenticator
gspread
htbuilder
scikit-image
sqlitecloud
matplotlib

psycopg2
supabase

fastapi[standard]
uvicorn
python-multipart

